pc_to_brainops/requirements.txt
mysql-connector-python
python-dotenv
psutil
